fix: MEDIUM bugs + DB migrations for remaining CRITICAL/HIGH
DB Migrations (059_add_stock_and_notification_tracking.sql):
- C1: Add stock_deducted column to order_items
- H2: Add notification_sent column to orders
- Update existing confirmed orders with proper values
- Add partial index for pending notifications

MEDIUM fixes (10/10):
- M1: Integer overflow fix in COUNT queries (shopFollowQueries, workerQueries)
- M2: Per-chain cache invalidation (cryptoPriceService)
- M3: Cleanup both subscriptions AND shop_subscribers tables
- M4: Optional chaining for scene.current.id (manageWallets)
- M5: Cleanup messageId in error handler (shopOnboarding)
- M6: Pass token to getShop (createFollow)
- M7: Handle "query too old" in answerCbQuery (paySubscription)
- M8: Max length validation for promoCode (chooseTier)
- M9: Move set() inside try block for finally cleanup (paymentSlice)
- M10: Check both AbortError and CanceledError (App.jsx)

+++ backend/src/models/shopFollowQueries.js
@@ -38,7 +38,7 @@ export const shopFollowQueries = {
         ss.owner_id as source_owner_id,
         u.username as source_username,
         (
-          SELECT COUNT(*)
+          SELECT COUNT(*)::int
           FROM synced_products sp
           JOIN products p_synced ON sp.synced_product_id = p_synced.id
           JOIN products p_source ON sp.source_product_id = p_source.id
@@ -47,7 +47,7 @@ export const shopFollowQueries = {
             AND p_synced.is_active = true
         ) as synced_products_count,
         (
-          SELECT COUNT(*)
+          SELECT COUNT(*)::int
           FROM products p
           WHERE p.shop_id = sf.source_shop_id
             AND p.is_active = true
@@ -77,7 +77,7 @@ export const shopFollowQueries = {
         ss.owner_id as source_owner_id,
         u.username as source_username,
         (
-          SELECT COUNT(*)
+          SELECT COUNT(*)::int
           FROM synced_products sp
           JOIN products p_synced ON sp.synced_product_id = p_synced.id
           JOIN products p_source ON sp.source_product_id = p_source.id
@@ -85,7 +85,7 @@ export const shopFollowQueries = {
             AND p_source.is_active = true
             AND p_synced.is_active = true
         ) as synced_products_count,
-        (SELECT COUNT(*) FROM products p WHERE p.shop_id = sf.source_shop_id AND p.is_active = true) as source_products_count
+        (SELECT COUNT(*)::int FROM products p WHERE p.shop_id = sf.source_shop_id AND p.is_active = true) as source_products_count
       FROM shop_follows sf
       JOIN shops ss ON sf.source_shop_id = ss.id
       LEFT JOIN users u ON ss.owner_id = u.id
